# Tidy up `Maxout`