--- a/src/deprecations.jl
+++ b/src/deprecations.jl
@@ -33,3 +33,7 @@ zeros(T::Type, dims...) = Base.zeros(T, dims...)
 
 ones32(::Type, dims...) = throw(ArgumentError("Flux.ones32 is always Float32, use Base.ones to specify the element type"))
 zeros32(::Type, dims...) = throw(ArgumentError("Flux.zeros32 is always Float32, use Base.zeros to specify the element type"))
+
+
+# v0.13 deprecations
+@deprecate Maxout(layers::Tuple) Maxout(layers...)