--- a/test/layers/basic.jl
+++ b/test/layers/basic.jl
@@ -109,13 +109,13 @@ import Flux: activations
     end
 
     @testset "simple alternatives" begin
-      mo = Maxout((x -> x, x -> 2x, x -> 0.5x))
+      mo = Maxout(x -> x, x -> 2x, x -> 0.5x)
       input = rand(40)
       @test mo(input) == 2*input
     end
 
     @testset "complex alternatives" begin
-      mo = Maxout((x -> [0.5; 0.1]*x, x -> [0.2; 0.7]*x))
+      mo = Maxout(x -> [0.5; 0.1]*x, x -> [0.2; 0.7]*x)
       input = [3.0 2.0]
       target = [0.5, 0.7].*input
       @test mo(input) == target