@mcabbott 2021-11-30
In fact, the improvement is less clear at larger sizes. But I think we can optimise the gradient of `max.(...)` more easily than pairwise `mapreduce`?
```
julia> m = Maxout(() -> Dense(50, 70, tanh), 5);

julia> x = rand(Float32, 50, 110);

julia> @btime gradient(sum∘m, $x);
  min 806.917 μs, mean 1.248 ms (1268 allocations, 1.61 MiB. GC mean 7.54%)   # before
  min 782.125 μs, mean 919.414 μs (751 allocations, 1.24 MiB. GC mean 6.52%)  # after

julia> @btime m3($x); # forwards
  min 303.500 μs, mean 353.118 μs (28 allocations, 422.41 KiB. GC mean 5.61%)  # before
  min 324.584 μs, mean 367.592 μs (22 allocations, 331.89 KiB. GC mean 4.31%)  # after
```