@ToucheSir 2021-12-01
The stranger result to me is the forward pass. It's not clear why the before version would be faster, and it appears both versions are unrolling the per-element comparisons.