@mcabbott 2021-12-01
I have been surprised before by `max`, I think it's very careful about NaN etc. Still seems strange here, though.
```
julia> x, y, z = (rand(100,100) for _ in 1:99);

julia> @btime $y .= max.($x, 0.0, $z);
  min 8.361 μs, mean 8.406 μs (0 allocations)

julia> @btime $y .= clamp.($x, clamp.(0.0, $z, Inf), Inf);  # just ifelse
  min 2.403 μs, mean 2.443 μs (0 allocations)
```