@ToucheSir 2021-12-02
Come to think of it though, `max.(...)` might not be such a bad idea for the GPU path. I don't have the time to run benchmarks at just this moment, but if that is the case we could make a case for eating the slightly slower forward pass on CPU for fewer allocations and a faster backwards one?