@mcabbott 2021-12-02
I was going to suggest reverting to `mapreduce` for now, this PR can just be minor tidying up & the next one can look more at performance.

I wrote some gradients for `max.(x,y,z)`, seems possible to make it roughly twice as quick on CPU. But it's more code than I thought it would be, how many high-performance hand-written kernels do we want in Zygote? 
https://gist.github.com/mcabbott/48aa0f4e0ec730d39995ad9917cad8f1

The last commit is unrelated, but makes CPU tests pass on 1.7. Could be cherry-picked if someone wants to make a separate PR. Have not investigated why the result is slightly different.